requirements.txt
fastapi
uvicorn[standard]
joblib
scikit-learn
pandas
paho-mqtt
asyncio-mqtt